Refactor isolating os-specific commands
